backend/pymes-plus-api/requirements.txt
alembic==0.8.6
coverage==4.2
dotenv==0.0.5
ecdsa==0.13
Flask==0.10.1
Flask-HTTPAuth==3.1.2
Flask-Testing==0.4.2
future==0.15.2
gunicorn==19.4.5
httpie==0.9.3
itsdangerous==0.24
Jinja2==2.8
m3-cdecimal==2.3
Mako==1.0.4
MarkupSafe==0.23
Pillow==3.3.0
pycrypto==2.6.1
pyDes==2.0.1
Pygments==2.1.3
PyJWT==1.4.0
PyMySQL==0.7.2
python-editor==1.0
python-jose==0.6.1
reportlab==3.3.0
requests==2.10.0
simplejson==3.8.2
six==1.10.0
SQLAlchemy==1.0.12
Werkzeug==0.11.9
gunicorn==19.4.5